refactor: :art: rename files to camelCase
