+++ README.md
@@ -3,11 +3,11 @@
 
 ## Hooks
 
-- [useArray](https://github.com/hdpinc/rehooks/tree/master/src/use-array)
-- [useBool](https://github.com/hdpinc/rehooks/tree/master/src/use-bool)
-- [useToggle](https://github.com/hdpinc/rehooks/tree/master/src/use-toggle)
-- [useClickAway](https://github.com/hdpinc/rehooks/tree/master/src/use-click-away)
-- [useKeyUp](https://github.com/hdpinc/rehooks/tree/master/src/use-key-up)
-- [useKeyDown](https://github.com/hdpinc/rehooks/tree/master/src/use-key-down)
-- [useDate](https://github.com/hdpinc/rehooks/tree/master/src/use-date)
-- [useDatepicker](https://github.com/hdpinc/rehooks/tree/master/src/use-datepicker)
+- [useArray](https://github.com/hdpinc/rehooks/tree/master/src/useArray)
+- [useBool](https://github.com/hdpinc/rehooks/tree/master/src/useBool)
+- [useToggle](https://github.com/hdpinc/rehooks/tree/master/src/useToggle)
+- [useClickAway](https://github.com/hdpinc/rehooks/tree/master/src/useClickAway)
+- [useKeyUp](https://github.com/hdpinc/rehooks/tree/master/src/useKeyUp)
+- [useKeyDown](https://github.com/hdpinc/rehooks/tree/master/src/useKeyDown)
+- [useDate](https://github.com/hdpinc/rehooks/tree/master/src/useDate)
+- [useDatepicker](https://github.com/hdpinc/rehooks/tree/master/src/useDatepicker)
