+++ src/index.ts
@@ -1,11 +1,11 @@
-import useArray from './use-array'
-import useBool from './use-bool'
-import useClickAway from './use-click-away'
-import useDate from './use-date'
-import useDatepicker from './use-datepicker'
-import useKeyDown from './use-key-down'
-import useKeyUp from './use-key-up'
-import useLocalStorage from './use-local-storage'
-import useToggle from './use-toggle'
+import useArray from './useArray'
+import useBool from './useBool'
+import useClickAway from './useClickAway'
+import useDate from './useDate'
+import useDatepicker from './useDatepicker'
+import useKeyDown from './useKeyDow'
+import useKeyUp from './useKeyUp'
+import useLocalStorage from './useLocalStorage'
+import useToggle from './useToggle'
 
 export { useArray, useBool, useClickAway, useDate, useDatepicker, useKeyDown, useKeyUp, useLocalStorage, useToggle }
