+++ src/useDatepicker/index.tsx
@@ -1,6 +1,6 @@
 import { format, isAfter, isBefore, isSameDay, isSameMonth, parseISO, startOfMonth } from 'date-fns'
 import React from 'react'
-import useDate, { DateLike, UseDateReturn } from '../use-date'
+import useDate, { DateLike, UseDateReturn } from '../useDate'
 import { getDateMatrixForYearMonth } from './utils'
 
 type InputMode = 'date' | 'month' | 'year'
